feat: add middleware

+++ routes/web.php
@@ -20,16 +20,16 @@ use App\Http\Controllers\AuthController;
 
 Route::get('/', [ContactController::class, 'displayUser']) ->name('home');
 
-Route::get('/contact/{id}', [ContactController::class, 'showUser']) ->name('contact.show_user');
-Route::get('/contact/{id}/edit', [ContactController::class, 'showEdit']) ->name('contact.edit');
-Route::put('/contact/{id}', [ContactController::class, 'update']) ->name('contact.update');
+Route::get('/contact/{id}', [ContactController::class, 'showUser']) ->name('contact.show_user')->middleware('auth');
+Route::get('/contact/{id}/edit', [ContactController::class, 'showEdit']) ->name('contact.edit')->middleware('auth');
+Route::put('/contact/{id}', [ContactController::class, 'update']) ->name('contact.update')->middleware('auth');
 
 
-Route::post('/post', [ContactController::class, 'addUser']) ->name('contact.add');
-Route::delete('/contact/{id}', [ContactController::class, 'deleteUser']) ->name('contact.destroy');
+Route::post('/post', [ContactController::class, 'addUser']) ->name('contact.add')->middleware('auth');
+Route::delete('/contact/{id}', [ContactController::class, 'deleteUser']) ->name('contact.destroy')->middleware('auth');
 
 
-Route::get('/register', [AuthController::class, 'register']) ;
+Route::get('/register', [AuthController::class, 'register']);
 Route::post('/register', [AuthController::class,'store'])->name('auth.register');
 
 Route::get('/login', [AuthController::class, 'login'])->name('login');
